feat(tae): zonemap with fixed 64-byte serialized data (#4348)
A zonemap with 64-byte serialized data.

If the data type is string, only a part of prefix of minimum and maximum will be written to disk
Rule of thumb: false positive is allowed but false negative is not
That means the searialized min-max range should cover the original min-max range.

Therefore, we must record minv length, because filling zero for minv makes it bigger, which is not acceptable.
For maxv, we have to construct a bigger value in 32 bytes by plus one if needed.
What if the leading 32 bytes are all 0xff? That is means +inf, it should be compared specifically

Layout for string:
```
[B0,...B30,B31,B32,...B62,B63]
---------  -  --------------
minv    |       maxv
|
[b7=init,b6~b5 unused,b4~b0=len(minv)]
```

Approved by: @XuPeng-SH