update buffer manager (#5265)
1. Change the node key from common.ID to any
2. Enhancement evict related logic
Previously, a buffer node needed to be registered first. When Pinning a node, it applies a size quota and loads data. When a node is evicted, it only unloads the data and returns back its size quota.
Now, when a hard evictable node is evicted, it will be destroyed and then removed from the buffer manager
3. Add PinByKey, TryPinByKey and Add APIs

Approved by: @yingfeng