temp unlimit the txn workspace cache size (#5326)
The txn workspace cache size limits the max cache usage used by all txns. If the total memory usage is greater than the limit, it will be spilled to the log file. During the txn commit, the data spilled into the log before will be loaded into the memory and applied to the state machine.

Since 0.6 we'll never deal with the big memory-consumed transactions on DN, we can set the limit to a reasonable value.

Approved by: @yingfeng