transfer deletes if the target block was compacted - phase 1 (#6490)
This conflict itself is very low frequency. At the current stage, I will try to reduce the frequency of this kind of conflict, but it will not prevent the occurrence of conflict. Only in some very extreme cases, may conflicts.

A very long transaction staged in CN and commit to DN later
The source block was compacted more than once
An update transaction first deletes and then appends a row with the same PK. If the compaction is committed in between, it may conflict. Given that DN applies deletes and appends in an extremely short interval, there is little chance of such a conflict.

Approved by: @LeftHandCold