Change the score of the flush block& NewBinaryFuseFilter deduplicate retry (#6486)
Change the score of the flush block and force the block to be compacted within a specified time. If a block is frequently deleted some rows regularly, the block may not be compacted.
After NewBinaryFuseFilter fails, you need to deduplicate the block and try again, because when the number of delete rows in a block exceeds 115, xorfilter will report an error due to too much duplicate data.

Approved by: @yingfeng, @XuPeng-SH, @fengttt