Change the cleaning method when the connection quits (#6499)
When the conneciton quits, in previous cleaning procedure, the transaction should be rollback. It is prone to trigger the concurrent problems.
Now, It sets a boolean label when the connection quits.
After the general procedure ends, it finds the boolean label and do the cleaning.
In this way, it prohibits rollbacking the transaction from reading/writing the transaction at the same time.

Approved by: @nnsgmsone, @yingfeng