fix bug: prepare with show stmt is incorrect (#7055)
fix bug: prepare with show stmt is incorrect.

some statement was execute before Compile. but we run execute stmt and get the origin stmt after Compile。
so we get some bug for the stmt which executed before compile.

but only some show/set stmt was supported in prepare now.  so I deal with these stmt in this PR.

when some day we support more DDL in prepare, we will no only  add code for prepare clause , but also need to  change these logic in frontend.

Approved by: @daviszhen, @aressu1985