Dealing with side case in relation to text type. (#7201)
The check of a big/small(more than 23) string should not use type.Width, and I use whether  " area == nil" instead.

The width of text is 0, as a sign of unlimited upper bound, And when compare with other string type, the result should
be like the width of text is always bigger.

Approved by: @ouyuanning, @heni02