Remove some unused bvt case (#7259)
the issue#3370 was created for just discussion several months ago.
our semantic to cast int to timestamp is diff to Mysql. So we cannot support the old bvt case from mysql. For example:
1. mysql: 20221008 ==> 2022.10.08
2. Mo: 20221008 ==> 20221008 seconds after 1970.
And cast '20221008' to timestamp, we have the same performance with mysql.

issue#3739:
old tag is confused.
we don't support avg / max / min for char and varchar type.
and there is no relevant needs for avg(date) and avg(datetime), so we removed it temporarily.

Approved by: @aressu1985