lockservice: add timestamp for lock operation (#8518)
A transaction uses commit timestamp as an argument when releasing a lock. This means that the transaction blocked by the transaction gets the commit timestamp, and in RC mode, the blocked transaction can use this timestamp + 1 as the latest timestamp for the RC query to get the latest data.

Approved by: @fengttt, @w-zr