[FIX] Added T_BlockId to all the type switch statements. (#8683)
The new DN vector is embedding the CN vector inside it. Hence we find the missing `T_BlockId` in some `type switch` to be a problem. `T_BlockId` was recently defined, and is only defined in a few `type switch` places.

In this PR we are adding T_BlockId to all the `type switch` statements (where ever type.RowID is defined). We have mimicked the `T_RowID` definition for `T_BlockId`.

Approved by: @aunjgr, @aptend, @XuPeng-SH, @daviszhen, @ouyuanning, @reusee, @nnsgmsone