moctl: add sync commit ts in mo cluster (#9104)
When we use push mode and turn on cn-no-wait, in a multi-CN test scenario, there may be multiple sessions that do not see the latest commit data between each other, resulting in many BVT cases failing randomly. To solve this problem, mo_ctl('cn', 'SyncCommit', '') is added to force the last commit timestamp of the whole cluster CNs to be synchronized. mo-tester can use this to fix the above problem.

Also, mo_ctl has been refactored to support mo_ctl to control any node in the MO cluster based on service id.