[Performance]: Predicate transitivity and join order [step 2] (#9311)
deduce new predicate, for example join on a=b and b=c, we deduce that a=c.
use new cond to explore best join order. but if this cond is not useful for new join order,
delete this cond to avoid performance loss.
tpch q5 and q9 will get better join order

Approved by: @aunjgr