txn: adjusting transaction wait-related logic (#9826)
At the time of transaction creation, a ts is obtained as the start timestamp of the transaction. To ensure that the complete data is visible at the start of the transaction, the logtail of all < snapshot ts is waited until it is applied when the transaction is created inside the txn client.

Inside the txn client, there is a waiter waiting for the LogTail to be applied, which will continuously receive the ts applied by the pushClient, and then the transaction will use the maximum applied LogTail ts currently received + 1 as the transaction's snapshot ts to ensure that the transaction can see the log tails corresponding to the max(applied log tail ts in txn client).

So here we need to use snapshotTS.Prev() to check logtail is applied

Approved by: @nnsgmsone