Fix select for update retry (#10102)
Let the lock operator support select for update. For select for update, the query cannot be executed until all the locks on the query rows are completed. Because in case of a conflict, RC mode requires seeing the latest result and retrying. If the query is executed before all locks are completed, the client will receive the wrong data

Approved by: @ouyuanning, @m-schen, @nnsgmsone