fix task not rescheduled after cn changed (#10215)
Previously, a task is regarded as expired if its runner is an expired cn.
However, expired CNs will be deleted in HAKeeper's state machine, so task scheduler cannot get expired CNs info.
Now a task is regarded as expired if its runner is not a working CN.

Approved by: @zhangxu19830126