fix logtail lagging gc (#10469)
related to issue #10440

Before this pr, every minute, dn checks if the max checkpoint is finished, if so, gc logtail using the end timestamp. However, it is possible that every check will find the checkpoint is unfinished, and skip gc logtail, leading to a lot of unused logtail in memory.

In this pr,  the start ts of the max checkpoint is used to calculate the end ts of the previous checkpoint.

Approved by: @XuPeng-SH